refactor: browser fonts utils
